# isme-nest-serrve.swagger.json
{
	"info": {
		"title": "isme-nest-serrve",
		"description": "",
		"version": "1.0.0"
	},
...
	"paths": {
...
		"/role/permissions/tree": {
			"get": {
				"summary": "角色权限树-by token",
				"deprecated": false,
				"description": "",
				"tags": ["权限"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"name": {
												"type": "string"
											},
											"code": {
												"type": "string"
											},
											"type": {
												"type": "string"
											},
											"parentId": {
												"type": "null"
											},
											"path": {
												"type": "string",
												"x-nullable": true
											},
											"redirect": {
												"type": "null"
											},
											"icon": {
												"type": "string"
											},
											"component": {
												"type": "string",
												"x-nullable": true
											},
											"layout": {
												"type": "string",
												"x-nullable": true
											},
											"keepAlive": {
												"type": "null"
											},
											"method": {
												"type": "null"
											},
											"description": {
												"type": "null"
											},
											"show": {
												"type": "boolean"
											},
											"enable": {
												"type": "boolean"
											},
											"order": {
												"type": "integer"
											},
											"children": {
												"type": "array",
												"items": {
													"type": "object",
													"properties": {
														"id": {
															"type": "integer"
														},
														"name": {
															"type": "string"
														},
														"code": {
															"type": "string"
														},
														"type": {
															"type": "string"
														},
														"parentId": {
															"type": "integer"
														},
														"path": {
															"type": "string"
														},
														"redirect": {
															"type": "null"
														},
														"icon": {
															"type": "string"
														},
														"component": {
															"type": "string"
														},
														"layout": {
															"type": "string",
															"x-nullable": true
														},
														"keepAlive": {
															"type": "boolean",
															"x-nullable": true
														},
														"method": {
															"type": "null"
														},
														"description": {
															"type": "null"
														},
														"show": {
															"type": "boolean"
														},
														"enable": {
															"type": "boolean"
														},
														"order": {
															"type": "integer"
														},
														"children": {
															"type": "array",
															"items": {
																"type": "object",
																"properties": {
																	"id": {
																		"type": "integer"
																	},
																	"name": {
																		"type": "string"
																	},
																	"code": {
																		"type": "string"
																	},
																	"type": {
																		"type": "string"
																	},
																	"parentId": {
																		"type": "integer"
																	},
																	"path": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"redirect": {
																		"type": "null"
																	},
																	"icon": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"component": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"layout": {
																		"type": "string",
																		"x-nullable": true
																	},
																	"keepAlive": {
																		"type": "null"
																	},
																	"method": {
																		"type": "null"
																	},
																	"description": {
																		"type": "null"
																	},
																	"show": {
																		"type": "boolean"
																	},
																	"enable": {
																		"type": "boolean"
																	},
																	"order": {
																		"type": "integer"
																	}
																},
																"required": ["id", "name", "code", "type", "parentId",
																	"path", "redirect", "icon", "component",
																	"layout", "keepAlive", "method", "description",
																	"show", "enable", "order"
																]
															}
														}
													},
													"required": ["id", "name", "code", "type", "parentId", "path",
														"redirect", "icon", "component", "layout", "keepAlive",
														"method", "description", "show", "enable", "order",
														"children"
													]
												}
											}
										},
										"required": ["id", "name", "code", "type", "parentId", "path", "redirect",
											"icon", "component", "layout", "keepAlive", "method", "description",
											"show", "enable", "order", "children"
										]
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
...
		"/role": {
			"post": {
				"summary": "新增角色",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			},
			"get": {
				"summary": "角色列表-all",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "array",
									"items": {
										"type": "object",
										"properties": {
											"id": {
												"type": "integer"
											},
											"code": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"enable": {
												"type": "boolean"
											}
										},
										"required": ["id", "code", "name", "enable"]
									}
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/role/page": {
			"get": {
				"summary": "角色列表-分页",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [{
					"name": "pageNo",
					"in": "query",
					"description": "",
					"required": false,
					"type": "integer",
					"x-example": 1
				}, {
					"name": "pageSize",
					"in": "query",
					"description": "",
					"required": false,
					"type": "integer",
					"x-example": 10
				}, {
					"name": "name",
					"in": "query",
					"description": "",
					"required": false,
					"type": "string",
					"x-example": "超级管理员"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {
								"code": {
									"type": "integer"
								},
								"message": {
									"type": "string"
								},
								"data": {
									"type": "object",
									"properties": {
										"pageData": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "integer"
													},
													"code": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"enable": {
														"type": "boolean"
													},
													"permissionIds": {
														"type": "array",
														"items": {
															"type": "integer"
														}
													}
												},
												"required": ["id", "code", "name", "enable", "permissionIds"]
											}
										},
										"total": {
											"type": "integer"
										}
									},
									"required": ["pageData", "total"]
								},
								"originUrl": {
									"type": "string"
								}
							},
							"required": ["code", "message", "data", "originUrl"]
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/role/{id}": {
			"delete": {
				"summary": "删除角色",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string",
					"x-example": "1"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			},
			"patch": {
				"summary": "修改角色",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string",
					"x-example": "1"
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"produces": ["application/json"]
			}
		},
		"/role/users/remove/{id}": {
			"patch": {
				"summary": "取消分配角色-批量",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string",
					"x-example": "1"
				}, {
					"name": "body",
					"in": "body",
					"schema": {
						"type": "object",
						"properties": {}
					}
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"consumes": ["application/json"],
				"produces": ["application/json"]
			}
		},
		"/role/users/add/{id}": {
			"patch": {
				"summary": "分配角色-批量",
				"deprecated": false,
				"description": "",
				"tags": ["角色"],
				"parameters": [{
					"name": "id",
					"in": "path",
					"description": "",
					"required": true,
					"type": "string",
					"x-example": "1"
				}, {
					"name": "body",
					"in": "body",
					"schema": {
						"type": "object",
						"properties": {}
					}
				}],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"schema": {
							"type": "object",
							"properties": {}
						}
					}
				},
				"security": [{
					"bearer": []
				}],
				"consumes": ["application/json"],
				"produces": ["application/json"]
			}
		}
	},
	"swagger": "2.0",
	"basePath": null,
...
	"securityDefinitions": {
		"bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
...
}